prov/efa: migrate rxr_pkt_type_req.c to efa_rdm_pke_req.c
rename rxr_pkt_type_req.c to efa_rdm_pke_req.c, and change
the "rxr_pkt" prefix to "efa_rdm_pke".

Moved some functions that operations on a packet type
to efa_rdm_pkt_type.h/c.
