+++ prov/efa/src/efa_hmem.c
@@ -32,7 +32,7 @@
 
 #include "efa.h"
 #include "efa_hmem.h"
-#include "rdm/rxr_pkt_type_req.h"
+#include "rdm/efa_rdm_pkt_type.h"
 
 #if HAVE_CUDA || HAVE_NEURON
 static size_t efa_max_eager_msg_size_with_largest_header(struct efa_domain *efa_domain) {
@@ -40,7 +40,7 @@ static size_t efa_max_eager_msg_size_with_largest_header(struct efa_domain *efa_
 
 	mtu_size = efa_domain->device->rdm_info->ep_attr->max_msg_size;
 
-	return mtu_size - rxr_pkt_max_hdr_size();
+	return mtu_size - efa_rdm_pkt_type_get_max_hdr_size();
 }
 #else
 static size_t efa_max_eager_msg_size_with_largest_header(struct efa_domain *efa_domain) {
