+++ prov/efa/src/rdm/efa_rdm_pke.c
@@ -48,7 +48,7 @@
 #include "efa_rdm_pke_cmd.h"
 #include "efa_rdm_pke_rtm.h"
 #include "efa_rdm_pke_nonreq.h"
-#include "rxr_pkt_type_req.h"
+#include "efa_rdm_pke_req.h"
 
 /**
  * @brief allocate a packet entry
